CIT-42 recolte Management

+++ src/main/java/com/citronix/citronix/repository/RecolteRepository.java
@@ -5,6 +5,6 @@ import com.citronix.citronix.entity.Enum.Saison;
 import com.citronix.citronix.entity.Recolte;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecolteRepository extends JpaRepository<Recolte, Integer> {
+public interface RecolteRepository extends JpaRepository<Recolte, Long> {
     boolean existsByChampAndSaison(Champ champ, Saison saison);
 }
